+++ p2panda-rs/src/document/document_view_id.rs
@@ -34,13 +34,33 @@ pub struct DocumentViewId(Vec<OperationId>);
 
 impl DocumentViewId {
     /// Create a new document view id.
-    pub fn new(graph_tips: Vec<OperationId>) -> Self {
-        Self(graph_tips)
+    pub fn new(graph_tips: &[OperationId]) -> Self {
+        Self(graph_tips.to_vec())
     }
 
     /// Get the graph tip ids of this view id.
     pub fn graph_tips(&self) -> &[OperationId] {
         self.0.as_slice()
+    }
+
+    /// Returns a hash over the sorted graph tips constituting this view id.
+    ///
+    /// Use this as a unique identifier for a document if you need a value with a limited size. The
+    /// document view id itself grows with the number of graph tips that the document has, which
+    /// may not be desirable for an identifier.
+    ///
+    /// Keep in mind that when you refer to document views with this hash value it will not be
+    /// possible to recover the document view id from it.
+    pub fn hash(&self) -> Hash {
+        // Sort graph tips to ensure consistent hashes
+        let mut graph_tips_mut = self.0.clone();
+        graph_tips_mut.sort();
+
+        let graph_tip_bytes = graph_tips_mut
+            .into_iter()
+            .flat_map(|graph_tip| graph_tip.as_hash().to_bytes())
+            .collect();
+        Hash::new_from_bytes(graph_tip_bytes).unwrap()
     }
 }
 
@@ -83,7 +103,7 @@ impl IntoIterator for DocumentViewId {
 /// only consists of one graph tip hash.
 impl From<OperationId> for DocumentViewId {
     fn from(operation_id: OperationId) -> Self {
-        Self::new(vec![operation_id])
+        Self::new(&[operation_id])
     }
 }
 
@@ -93,7 +113,7 @@ impl From<OperationId> for DocumentViewId {
 /// consists of one graph tip hash.
 impl From<Hash> for DocumentViewId {
     fn from(hash: Hash) -> Self {
-        Self::new(vec![hash.into()])
+        Self::new(&[hash.into()])
     }
 }
 
@@ -105,7 +125,7 @@ impl FromStr for DocumentViewId {
     type Err = HashError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self::new(vec![Hash::new(s)?.into()]))
+        Ok(Self::new(&[Hash::new(s)?.into()]))
     }
 }
 
@@ -116,7 +136,7 @@ mod tests {
     use crate::hash::Hash;
     use crate::operation::OperationId;
     use crate::test_utils::constants::DEFAULT_HASH;
-    use crate::test_utils::fixtures::{document_view_id, random_hash};
+    use crate::test_utils::fixtures::{document_view_id, random_hash, random_operation_id};
     use crate::Validate;
 
     use super::DocumentViewId;
@@ -128,16 +148,16 @@ mod tests {
         let document_id: DocumentViewId = hash_str.parse().unwrap();
         assert_eq!(
             document_id,
-            DocumentViewId::new(vec![hash_str.parse::<OperationId>().unwrap()])
+            DocumentViewId::new(&[hash_str.parse::<OperationId>().unwrap()])
         );
 
         // Converts a `Hash` to `DocumentViewId`
         let document_id: DocumentViewId = hash.clone().into();
-        assert_eq!(document_id, DocumentViewId::new(vec![hash.clone().into()]));
+        assert_eq!(document_id, DocumentViewId::new(&[hash.clone().into()]));
 
         // Converts an `OperationId` to `DocumentViewId`
         let document_id: DocumentViewId = OperationId::new(hash.clone()).into();
-        assert_eq!(document_id, DocumentViewId::new(vec![hash.into()]));
+        assert_eq!(document_id, DocumentViewId::new(&[hash.into()]));
 
         // Fails when string is not a hash
         assert!("This is not a hash".parse::<DocumentViewId>().is_err());
@@ -171,5 +191,17 @@ mod tests {
             format!("{}", view_id_unmerged),
             "0020b177ec1bf26dfb3b7010d473e6d44713b29b765b99c6e60ecbfae742de496543_0020d3235c8fe6f58608200851b83cd8482808eb81e4c6b4b17805bba57da9f16e79"
         );
+    }
+  
+    #[rstest]
+    fn document_view_hash(
+        #[from(random_operation_id)] operation_id_1: OperationId,
+        #[from(random_operation_id)] operation_id_2: OperationId,
+    ) {
+        let view_id_1 = DocumentViewId::new(&[operation_id_1.clone(), operation_id_2.clone()]);
+        assert!(view_id_1.hash().validate().is_ok());
+
+        let view_id_2 = DocumentViewId::new(&[operation_id_2, operation_id_1]);
+        assert_eq!(view_id_1.hash(), view_id_2.hash());
     }
 }